Remove a need for compilation step (jsx)
